Fix incorrect usage of rolebinding vs clusterrolebinding (#55)
I had used a RoleBinding even though the role that was being referenced was a ClusterRole. This isn't allowed. Though why the hell they are different in any way I don't really fucking know. So here we go.

Thanks for that awesome http method not allowed error message. It's so great.