Remove the half baked dark colors from static (#1321)
Summary:
Dark colors are a nice to have. However they only work if we have the rest of the color scheme to back it up. As it was these colors were making light grey text on white backgrounds all over the place.

Test Plan:
Let the netlify bot chew on this