feat: annotate kube resource with configmap and secrets mounted as volumes (#1495)
Summary:
One of the most common thing with kube services is that they don't restart when the files from a volume change. Ory keto is very bad at this. So this add the feature to add those tags.

Test Plan:
Got it working and saw that istio had some tags