feat: retry failed snapshot apply (#1740)
Summary:

Some Kubernetes operators register CRD's after they start up, or the
CRD's require webhooks to be up and healthy to use. Since we start the
operators and try syncing all resources at the same time this can lead
to failures. This pr adds a new GenServer that will schedule a new
snapshot apply in the future when noticing a failed apply.

- New Genserver
- Add the genserver to the `kube_services` application supervision tree.

Test Plan:

Made the code change.
Started a clean cluster
Waited for the first sync and the inital postgres restart
Then added the MetalLB battery
Notice the error result
Notice that the deploys got retried.