refactor: remove kube_ext (#1767)
Summary:
It was getting unclear if things should be in kube_ext or kube_services.
Likewise for kube_resource or kube_ext. So this removes kube_ext.
Anything that's a genserver or starts a process that needs a kubernetes,
is in the kube_services application.

Anything that's about resources is in the kube_resources application.

Test Plan:
`mix do clean, compile --force --warnings-as-errors`