feat: use local static spec for bootstrap (#2355)
Summary:
Previously we were download the install spec for bootstrapping, because
that is what will happen in production from home base. However that
poses challenges in long lived branches that make changes to install
spec.

This changes the cli to accept a new param static-dir that will take in
where the static directory is. Then we read that file if it's there.

Test Plan:

`bi bootstrap` shows the local file being read.