fix: overwrite xdg state for installs on fetching spec (#3339)
Summary:
Some steps of install add to the install spec. So we need to be careful
about overwriting with less information.

This PR simplifies things. Whenever we fetch an install spec, we erase
everything else and start over. That means that start will add things
and then write. Nothing should change the files after that.

However bix bootstrap takes in a file path. So this adds a command to
get the slug for an install spec. Then we use that everywhere.

Test Plan:
bix bootstrap.

CI