fix: get keycloak re-working (#82)
The schema changes broke a bunch of keycloak stuff. The primary problem
was that `insert_all` only autogenerates binary and int IDs. When they
changed to battery UUIDs, they were no longer being generated. It also
caused the client IDs to be too long with the `batt_` prefix.

It can now start up again and SSO works for both native and auth proxy
enabled services.

I want to take a quick pass and see if I can add some tests to prevent
future regressions like this as a separate PR.

---------

Co-authored-by: Jason Tarasovic <2999202+JTarasovic@users.noreply.github.com>