Fix the way switch inputs work with boolean values (#189)
This PR fixes the way switch inputs handle boolean inputs. Rather than
passing an empty value to the input tag if nothing is defined (which
makes the browser default to "on" and "off"), it uses the same hidden
input trick as the checkbox with "true" and "false". Requires the usage
of
[Phoenix.HTML.Form.normalize_value/2](https://hexdocs.pm/phoenix_html/Phoenix.HTML.Form.html#normalize_value/2)
when using the raw form value anywhere else.

[See this discussion in
Slack.](https://batteries-included.slack.com/archives/C077UVAP13M/p1718380577027399)