Create InputList component (#807)
This PR adds a new `InputList` component for showing a list of inputs
from an Ecto field that can be easily added to/removed from, and uses it
for the AWS battery config and the new team form in home base.

There are essentially two different variants since they need to be
handled differently: one for association fields such as
`has_many`/`embeds_many`, and one for basic array fields such as
`{:array, :string}`. There is still a bug where if there are two basic
array fields in a form and you add an item for one of the fields, the
other field will show an empty field that can't be removed until the
form is saved. I've been pulling my hair out trying to figure this out
and have wasted enough time on it, so I will create an issue to fix this
later since it's not a huge deal.

## Other Changes

- Remove extra whitespace when an input has no label
- Fix `ACTIVE` badges for batteries in dark mode

---

<img width="483" alt="Screenshot 2024-09-11 at 11 30 34"
src="https://github.com/user-attachments/assets/5a237ca9-da99-477e-960c-524c57c27577">
<img width="481" alt="Screenshot 2024-09-11 at 11 31 00"
src="https://github.com/user-attachments/assets/608832f1-dc3b-4087-aa20-91953c8987d7">