Add a mix task for adding an admin to internal teams (#833)
For https://github.com/batteries-included/batteries-included/issues/828.
This lets you run `mix home_base.admin jane@doe.com` to add a user
account to the internal admin teams.

I'm not sure what the best way is to start the home base app so we can
have access to the database. This seems to work with
`Mix.Task.run_in_apps/2`, but it overwhelms the logs with a bunch of
supervisor stuff, swallowing the logger messages from the task.