security: remove all traces of the private key as early as possible (#914)
Summary:
For every install we generate a curve pair. Home base will verify all
messages for usage and hosts with that key. We don't want others to get
it. So this makes the first report to remove the key.

This also adds some palces where we assert that the keys are using our
curve. That's becuase I made assumptions about how the keys are
structured based upon this key type.

Test Plan:
- Tests Added