Deprecate labels, help text, and notes for Input (#1090)
This simplifies the Input component by deprecating the label, help text,
and note attributes. I deprecated instead of removed, since there are
still tons of inputs that use those attributes. Once all those inputs
are updated to use the Fieldset component, the deprecated attributes and
functions can be removed.

## Other Changes

- Removes title attribute from Fieldset component (not needed)
- Fixes the structure of the Field component so the note is not part of
the label
- Fixes minor layout bugs with some of the input types
- Adds a story for the password input type

Related to
https://github.com/batteries-included/batteries-included/issues/23