ci: don't push latest image on tagging/release (#1095)
We can tag a commit at any point. Meaning that there could be more
recent commits so we shouldn't tag the released image as "latest". This
accomplishes that by determining which tag to use and then only running
`push-images` once with the tags that it should use as opposed to
running once with latest and once with the current tag.

Co-authored-by: Jason Tarasovic <2999202+JTarasovic@users.noreply.github.com>