ci: fix random naming issue (#1974)
I figured out why we were having issues with naming.
We auto-generate and re-set the names for e.g. postgres clusters on
validation.
When wallaby clears and sets the value, it causes phoenix to reset the
name in between clearing and setting. We can use set_value to send 100
backspaces and then the name we want and it works just fine!

This is needed to e.g. pick the postgres cluster for ferretdb.

Co-authored-by: Jason Tarasovic <2999202+JTarasovic@users.noreply.github.com>